feat(auth): added utils to get and check login header
