+++ src/auth/index.ts
@@ -1,3 +1,4 @@
 export * from './AuthManager'
 export * from './isAuthorizeFormRequired'
 export * from './isLoginRequired'
+export * from './loginHeader'
