+++ src/auth/spec/mockResponses.ts
@@ -1,8 +1,6 @@
 import { SasAuthResponse } from '@sasjs/utils/types'
-import { enLoginSuccessHeader } from '../AuthManager'
 
 export const mockLoginAuthoriseRequiredResponse = `<form id="application_authorization" action="/SASLogon/oauth/authorize" method="POST"><input type="hidden" name="X-Uaa-Csrf" value="2nfuxIn6WaOURWL7tzTXCe"/>`
-export const mockLoginSuccessResponse = enLoginSuccessHeader
 
 export const mockAuthResponse: SasAuthResponse = {
   access_token: 'acc355',
