+++ src/auth/spec/verifySas9Login.spec.ts
@@ -3,7 +3,7 @@
  */
 import { verifySas9Login } from '../verifySas9Login'
 import * as delayModule from '../../utils/delay'
-import { enLoginSuccessHeader } from '../AuthManager'
+import { getExpectedLogInSuccessHeader } from '../'
 
 describe('verifySas9Login', () => {
   const serverUrl = 'http://test-server.com'
@@ -19,7 +19,9 @@ describe('verifySas9Login', () => {
     const popup = {
       window: {
         location: { href: serverUrl + `/SASLogon` },
-        document: { body: { innerText: `<h3>${enLoginSuccessHeader}</h3>` } }
+        document: {
+          body: { innerText: `<h3>${getExpectedLogInSuccessHeader()}</h3>` }
+        }
       }
     } as unknown as Window
