+++ src/auth/verifySas9Login.ts
@@ -1,5 +1,5 @@
 import { delay } from '../utils'
-import { enLoginSuccessHeader } from './AuthManager'
+import { getExpectedLogInSuccessHeader } from './'
 
 export async function verifySas9Login(loginPopup: Window): Promise<{
   isLoggedIn: boolean
@@ -7,13 +7,17 @@ export async function verifySas9Login(loginPopup: Window): Promise<{
   let isLoggedIn = false
   let startTime = new Date()
   let elapsedSeconds = 0
+
   do {
     await delay(1000)
     if (loginPopup.closed) break
 
     isLoggedIn =
       loginPopup.window.location.href.includes('SASLogon') &&
-      loginPopup.window.document.body.innerText.includes(enLoginSuccessHeader)
+      loginPopup.window.document.body.innerText.includes(
+        getExpectedLogInSuccessHeader()
+      )
+
     elapsedSeconds = (new Date().valueOf() - startTime.valueOf()) / 1000
   } while (!isLoggedIn && elapsedSeconds < 5 * 60)
